Make the Javascript client understand the new JSON format.
When a log is named in the hash, the client now loads a JSON blob
from the server. And the client now uploads both the parsed JSON data
and the raw text file.

To do: make the server handle the JSON blob smarter. Goal is to split
the raw data out from the parsed data and store in separate files. The
raw data conceivably contains private info.